[BNNS] Min. Rename file
